requirements.txt
Django==2.2.4
djangorestframework==3.10.1
gunicorn==19.9.0
django-filter==2.2.0